// src/generated/i18n-types.ts
export const I18N_KEYS = [
  "availableThemesAriaLabel",
  "buttonGuideSubtitle",
  "buttonGuideTitle",
  "captionsButtonLabel",
  "captionsDisableAriaLabel",
  "captionsDisableTitle",
  "captionsEnableAriaLabel",
  "captionsEnableTitle",
  "captionsUnavailableTitle",
  "controlBarSectionCaptions",
  "controlBarSectionMusic",
  "controlBarSectionPin",
  "controlBarSectionQuality",
  "controlBarSectionSpeed",
  "controlBarSectionTime",
  "controlBarSectionVolume",
  "controlBarSectionsAriaLabel",
  "controlBarSectionsSubtitle",
  "controlBarSectionsTitle",
  "extActionTitle",
  "extDescription",
  "extName",
  "guideCaptionsLeftClick",
  "guideCaptionsRightClick",
  "guideCategoryCaptions",
  "guideCategoryCurrentSpeed",
  "guideCategoryCurrentVolume",
  "guideCategoryMusic",
  "guideCategoryResolution",
  "guideCategorySpeed",
  "guideCategoryTime",
  "guideCategoryVolume",
  "guideCurrentSpeedLeftClick",
  "guideCurrentSpeedRightClick",
  "guideCurrentVolumeLeftClick",
  "guideCurrentVolumeRightClick",
  "guideMusicLeftClick",
  "guideMusicRightClick",
  "guideResolutionLeftClick",
  "guideResolutionRightClick",
  "guideSpeedLeftClick",
  "guideSpeedRightClick",
  "guideTimeLeftClick",
  "guideTimeRightClick",
  "guideVolumeLeftClick",
  "guideVolumeRightClick",
  "leftClickLabel",
  "musicModeDisableAriaLabel",
  "musicModeDisableTitle",
  "musicModeEnableAriaLabel",
  "musicModeEnableTitle",
  "pinDisableAriaLabel",
  "pinDisableTitle",
  "pinEnableAriaLabel",
  "pinEnableTitle",
  "popupAppearanceToggleToDarkAriaLabel",
  "popupAppearanceToggleToLightAriaLabel",
  "popupDocumentTitle",
  "popupHeaderSubtitle",
  "popupHeaderTitle",
  "previewAriaLabel",
  "qualityAutoLabel",
  "qualityAutoTitle",
  "qualityResolutionTitle",
  "rateResetAriaLabel",
  "rateResetTitle",
  "resetButton",
  "resetSectionsButton",
  "rightClickLabel",
  "seekForward15Label",
  "seekForward15Title",
  "seekForward30Label",
  "seekForward30Title",
  "seekForward60Label",
  "seekForward60Title",
  "seekRewind15Label",
  "seekRewind15Title",
  "seekRewind30Label",
  "seekRewind30Title",
  "seekRewind60Label",
  "seekRewind60Title",
  "speedDecreaseLabel",
  "speedDecreaseTitle",
  "speedIncreaseLabel",
  "speedIncreaseTitle",
  "themeClassicDescription",
  "themeClassicName",
  "themeDarkDescription",
  "themeDarkName",
  "themeOceanDescription",
  "themeOceanName",
  "themeRoseDescription",
  "themeRoseName",
  "themeYoutubeDescription",
  "themeYoutubeName",
  "themesSubtitle",
  "themesTitle",
  "volumeDecreaseLabel",
  "volumeDecreaseTitle",
  "volumeIncreaseLabel",
  "volumeIncreaseTitle"
] as const;

export type I18nKey = (typeof I18N_KEYS)[number];